+++ packages/account-react/src/wagmi/wagmi-connector.tsx
@@ -1,4 +1,4 @@
-import { sophonSsoConnector } from '@sophon-labs/account-connector';
+import { createSophonConnector } from '@sophon-labs/account-connector';
 import { useEffect, useMemo } from 'react';
 import { useConfig } from 'wagmi';
 import { useSophonContext } from '../hooks/useSophonContext';
@@ -12,7 +12,7 @@ export const SophonWagmiConnector = ({
   const { partnerId, authServerUrl, network, updateConnector, communicator } =
     useSophonContext();
   const connector = useMemo(() => {
-    return sophonSsoConnector(partnerId, network, {
+    return createSophonConnector(partnerId, network, {
       authServerUrl,
       communicator,
     });