+++ packages/account-connector/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-connector
 
+## 2.0.10-next.0
+
+### Patch Changes
+
+- Exposing sophon connector and sophon provider helper utility
+- Updated dependencies
+  - @sophon-labs/account-core@2.0.10-next.0
+
 ## 2.0.9
 
 ### Patch Changes